🎨 rename folder from preferences to preference
