+++ backend/src/app.module.ts
@@ -6,7 +6,7 @@ import { RegisterModule } from './register/register.module';
 import { LoginService } from './login/login.service';
 import { LoginController } from './login/login.controller';
 import { LoginModule } from './login/login.module';
-import { PreferenceModule } from './preferences/preference.module';
+import { PreferenceModule } from './preference/preference.module';
 @Module({
   imports: [
     TypeOrmModule.forRoot({
