+++ backend/src/register/register.module.ts
@@ -5,7 +5,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRepository } from './repositories/user.repository';
 import { TrainerRepository } from './repositories/trainer.repository';
 import { TraineeRepository } from './repositories/trainee.repository';
-import { Preference } from '../preferences/entities/preference.entity';
+import { Preference } from '../preference/entities/preference.entity';
 @Module({
   imports: [
     TypeOrmModule.forFeature([
